ajout de dismit

+++ lib/src/screens/project_screen/project_item_builder.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:projects3/src/daos/auth.dart';
 import 'package:projects3/src/daos/project_dao.dart';
 import 'package:projects3/src/models/project.dart';
@@ -16,20 +17,48 @@ class ProjectItemBuilder extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return 
-    Dismissible(
-                    key: ObjectKey(project.id),
-                    background: Container(color: Colors.red),
-                    onDismissed: (direction) {
-                     // documentSnapshot.data().remove(index);
-                     print(project.titre);
-                     ProjectDao.deleteP(project.id, Auth.uid);
-                     
-                     
-                    },
-                    
+Dismissible(
+    background: Container(
+      alignment: Alignment.centerLeft,
+      padding: EdgeInsets.only(left: 20.0),
+      color: Colors.orangeAccent,
+      child: Icon(Icons.edit, color: Colors.white),
+    ),
+    secondaryBackground: Container(
+      alignment: Alignment.centerRight,
+      padding: EdgeInsets.only(right: 20.0),
+      color: Colors.redAccent,
+      child: Icon(Icons.delete, color: Colors.white),
+    ),
+    child: ListTile(
+      title: Text(project.titre)  ,
+      subtitle: Text(
+        "date fin: ${project.dateFin.toString()}"
+        
+        ),
+        
+        trailing:  Text(project.dateDedut.toString()),
+      onTap: (){
+        ontap(project: project,selectedScreen:ListTache.screenName);
+      },
+      
+    ),
+    key: ObjectKey(project.id),
+    onDismissed: (direction){
+      if(direction == DismissDirection.startToEnd){
+        print(project.id);
+        
+      } else if(direction == DismissDirection.endToStart){
+        //Scaffold.of(context).showSnackBar(SnackBar(content: Text("Swipe to right")));
+        print(project.id);
+        print("supprimer");
+        ProjectDao.deleteP(project.id, Auth.uid);
+      }
+    },
 
-                    confirmDismiss: (DismissDirection direction) async {
-  return await showDialog(
+    confirmDismiss: (direction) async {
+      if(direction==DismissDirection.endToStart){
+ return await showDialog(
     context: context,
     builder: (BuildContext context) {
       return AlertDialog(
@@ -47,29 +76,19 @@ class ProjectItemBuilder extends StatelessWidget {
           ),
         ],
       );
-    },
-  );
+      }     
+ );
+
+
+    }
+
+
+    else{
+      print("modifier");
+    }
+  
+  
 },
-
-                    
-
-                    direction: DismissDirection.endToStart,
-    child: Card(
-   child:
-    
-    ListTile(
-      title: Text(project.titre)  ,
-      subtitle: Text(
-        "date fin: ${project.dateFin.toString()}"
-        
-        ),
-        
-        trailing:  Text(project.dateDedut.toString()),
-      onTap: (){
-        ontap(project: project,selectedScreen:ListTache.screenName);
-      },
-    )
-    )
-    );
+  );
   }
 }
